Add control microbenchmarks.
Add stubbed out versions of the microbenchmarks so we can see various
overheads. I tried to make sure they generate the right code and don't
bypass the test entirely, but I still need to check the codegen to be
sure about it. Still, it shows the local versions as being about 50x
faster.