Add a third latch/phase to the benchmark.
I realized that the main thread would race with thread startup to start
the timers early, then block on starting the test. Adding a third phase
fixes that, and makes the "stub" benchmarks much faster. It makes them
go from a cpu time of ~26k to ~16k, and from 15-22 ms to 8-11 ms of wall
time, which is much more consistent.

Gotta keep lowering that noise floor.