Fix warning for value/reference.
Yeah, GCC warns on this despite how cheap it is to just copy it. I don't
know an idomatic way to surpress it, so just do like it says (and like
is usually best) and take it by reference.

Also I don't know, maybe it'll optimize out the extra level of
indirection? Whatever, it's not the point of the benchmark anyways.