Refactor test/benchmark helpers into another file.
The test and benchmark programs had some code in common, so this
refactors that out. Yeah, it means the testing function has a little
more overhead, but that's okay, it doesn't need to be fast anyways.

Also fuss with the makefile a bit since I was having trouble with
something that turned out to be a typo. Dropping the "./" from the
include paths is probably a good idea, though.