Add a MPMC queue that uses semaphores
Add a MPMC queue class that uses semaphores to keep track of how many
items are available in the queue. This replaces the condition_variable,
and potentially avoids a few lock/unlock actions on a mutex.

Unfortunately, I can't figure out a way to ask for LeastMaxValue to be
as big as possible, so I have to do use int, because that's what
libstdc++'s counting_semaphore is willing to promise.